# cloud/api/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "SuperRoo Cloud API",
		"description": "RESTful API for the SuperRoo Cloud Dashboard — job queue management, agent orchestration, Telegram bot integration, ML engine sync, multi-tenant organizations, monitoring, and IDE workspace management.\n\nBase URL: http://<host>:8787\n\nAll endpoints are also accessible with an /api prefix (e.g., /api/health).\n\n## Authentication\n\nMost endpoints require a Bearer token obtained from POST /auth/login or POST /auth/register.\n\n```\nAuthorization: Bearer <session_token>\n```\n\n## WebSocket Endpoints\n\n- `/api/ws/chat?sessionId=<id>` — AI Chat\n- `/api/ws/lsp?lang=<language>` — LSP Bridge (IntelliSense)\n- `/api/ws/telegram` — Telegram Bot real-time events\n- `/api/ws/dashboard` — Dashboard real-time updates",
		"version": "1.0.0",
		"contact": {
			"name": "SuperRoo Team",
			"url": "https://github.com/superroo"
		}
	},
	"servers": [
		{
			"url": "http://localhost:8787",
			"description": "Local development"
		},
		{
			"url": "http://104.248.225.250:8787",
			"description": "Production VPS"
		}
	],
...
	"paths": {
...
		"/monitoring/logs": {
			"get": {
				"tags": ["Monitoring"],
				"summary": "Get monitoring logs",
				"operationId": "getMonitoringLogs",
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 50
						}
					},
					{
						"name": "level",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["info", "warn", "error"]
						}
					},
					{
						"name": "source",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "search",
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Log entries"
					}
				}
			}
		},
...
		"/monitoring/alerts": {
			"get": {
				"tags": ["Monitoring"],
				"summary": "Get alert history",
				"operationId": "getAlertHistory",
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 100
						}
					},
					{
						"name": "offset",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 0
						}
					}
				],
				"responses": {
					"200": {
						"description": "Alert history"
					}
				}
			}
		},
...
		"/monitoring/alerts/{id}/ack": {
			"post": {
				"tags": ["Monitoring"],
				"summary": "Acknowledge alert",
				"operationId": "acknowledgeAlert",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "Alert ID"
					}
				],
				"responses": {
					"200": {
						"description": "Alert acknowledged"
					}
				}
			}
		},
		"/monitoring/alerts/{id}/resolve": {
			"post": {
				"tags": ["Monitoring"],
				"summary": "Resolve alert",
				"operationId": "resolveAlert",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "Alert ID"
					}
				],
				"responses": {
					"200": {
						"description": "Alert resolved"
					}
				}
			}
		},
...
	},
...
}